:boom: rewritten createMessage functions
