+++ lib/createMessage/fromMessenger.discord.js
@@ -1,8 +1,9 @@
 const log = require('npmlog')
+const handleMentions = require('../discord/handleMentions')
+const resolveImageURL = require('./resolveImageURL')
+const emojiCount = require('./emojiCount')
 const url = require('url')
-const handleMentions = require('./handleMentions')
 
-// function creating message from template
 module.exports = (thread, sender, message) => {
   var attach = message.attachments[0]
 
@@ -47,30 +48,4 @@ module.exports = (thread, sender, message) => {
     return true
   }).map(resolveImageURL).filter(a => a)
   return [ discordMessage, opts ]
-}
-
-function resolveImageURL (attach) {
-  var baseURL = attach.image || attach.url
-  var parse = u => url.parse(u, true)
-  if (!baseURL) return
-  baseURL = baseURL.match(/^(http|https):\/\/l\.facebook\.com\/l\.php/i) ? parse(baseURL).query.u : baseURL
-  var imageURL = parse(baseURL).pathname === '/safe_image.php' ? parse(baseURL).query.url : baseURL
-  var name = parse(imageURL).pathname.split('/').pop('-1')
-  return {
-    name: name.includes('.') ? name : name + '.png',
-    attachment: imageURL
-  }
-}
-
-// https://blog.jonnew.com/posts/poo-dot-length-equals-two
-function emojiCount (str) {
-  const split = str.split('\u{200D}')
-  let count = 0
-
-  for (const s of split) {
-    const num = Array.from(s.split(/[\ufe00-\ufe0f]/).join('')).length
-    count += num
-  }
-
-  return count / split.length
 }
