+++ lib/createMessage/getStreamFromURL.js
@@ -1,0 +1,5 @@
+const url = require('url')
+
+module.exports = attachURL => {
+  return new Promise((resolve, reject) => require(url.parse(attachURL).protocol.replace(':', '')).get(attachURL, res => resolve(res)))
+}
