+++ lib/discord/index.js
@@ -1,5 +1,4 @@
 module.exports = {
-  createMessage: require('./createMessage'),
   login: require('./login'),
   sendMessage: require('./sendMessage')
 }
