+++ lib/discord/listener.js
@@ -1,7 +1,6 @@
 const log = require('npmlog')
-const url = require('url')
 
-const { createMessage } = require('../messenger')
+const createMessage = require('../createMessage').fromDiscord
 const handleCommand = require('../handleCommand')
 const sendMessage = require('./sendMessage')
 
@@ -24,29 +23,12 @@ module.exports = async message => {
   // make sure this channel is meant for the bot
   if (!connections.has(message.channel.id)) return log.verbose('discordListener', 'Channel not found in bot\'s channel map')
 
-  // get image url from discord embeds
-  var imageUrl = message.embeds.length > 0
-    ? (message.embeds[0].image
-      ? message.embeds[0].image.url
-      : (message.embeds[0].thumbnail
-        ? message.embeds[0].thumbnail.url
-        : undefined
-      )
-    )
-    : undefined
-
-  // build message with attachments provided
-  const msg = {
-    body: createMessage(message),
-    attachment: await Promise.all([imageUrl].concat(message.attachments.map(attach => attach.url)).filter(el => el).map(getStreamFromURL))
-  }
-  log.silly('discordListener: message', msg)
   // find threads by channel ID
   const threads = connections.getThreads(message.channel.id).filter(el => !el.readonly)
   log.verbose('discordListener: threads', threads)
 
   // send message to threads specified in the config/channel topic
-  threads.forEach(threadID => messenger.client.sendMessage(msg, threadID, sentMessageCallback))
+  threads.forEach(async thread => messenger.client.sendMessage(await createMessage.messenger(message), thread.id, sentMessageCallback))
 
   const channels = (await connections.getChannels(message.channel.id)).filter(el => el.id !== message.channel.id)
   if (channels.length) {
@@ -60,10 +42,6 @@ module.exports = async message => {
     ]
     channels.forEach(channel => sendMessage(channel, message.channel.name, m))
   }
-}
-
-function getStreamFromURL (attachURL) {
-  return new Promise((resolve, reject) => require(url.parse(attachURL).protocol.replace(':', '')).get(attachURL, res => resolve(res)))
 }
 
 function sentMessageCallback (err, info) {
