+++ lib/messenger/index.js
@@ -1,5 +1,4 @@
 module.exports = {
-  createMessage: require('./createMessage'),
   filter: require('./filter'),
   getChannelName: require('./getChannelName'),
   getSender: require('./getSender'),
