+++ lib/messenger/listener.js
@@ -1,8 +1,8 @@
 const log = require('npmlog')
-const url = require('url')
 
-const { createMessage, sendMessage } = require('../discord')
-const { createMessage: createFBMessage, filter, getChannelName, getSender, getThread } = require('./')
+const createMessage = require('../createMessage').fromMessenger
+const { sendMessage } = require('../discord')
+const { filter, getChannelName, getSender, getThread } = require('./')
 const handleEvent = require('./handleEvent')
 var reconnecting
 
@@ -56,32 +56,18 @@ module.exports = async (error, message) => {
   // handle white/blacklist
   if (!filter(cleanname, message.threadID)) return
 
-  // build message from template
-  var m = createMessage(thread, sender, message)
-  log.silly('messengerListener: pending message', m)
-
   // get channel
   const channels = await connections.getChannels(message.threadID, cleanname)
   if (!channels || !channels.length) return
   log.verbose('messengerListener', 'Got Discord channels')
 
-  const sentMessages = await Promise.all(channels.map(channel => sendMessage(channel, cleanname, m, thread.imageSrc)))
+  await Promise.all(channels.map(channel => sendMessage(channel, cleanname, createMessage.discord(thread, sender, message), thread.imageSrc)))
 
   // check if it needs resending (linked channels)
-  const threads = connections.getThreads(message.threadID).filter(id => id !== message.threadID)
+  const threads = connections.getThreads(message.threadID).filter(thread => thread.id !== message.threadID).filter(el => !el.readonly)
   if (threads.length) {
-    const nickname = thread.nicknames[message.senderID]
-    const authorName = nickname ? (config.discord.showFullNames ? `${nickname} (${sender.name})` : nickname) : sender.name
-    const msg = {
-      body: createFBMessage(sentMessages[0].content, authorName, await getChannelName(thread, false)),
-      attachment: await Promise.all(sentMessages[0].attachments.map(attach => attach.url).filter(el => el).map(getStreamFromURL))
-    }
-    threads.forEach(threadID => messenger.client.sendMessage(msg, threadID, sentMessageCallback))
+    threads.forEach(async _thread => messenger.client.sendMessage(await createMessage.messenger(thread, sender, message, cleanname), _thread.id, sentMessageCallback))
   }
-}
-
-function getStreamFromURL (attachURL) {
-  return new Promise((resolve, reject) => require(url.parse(attachURL).protocol.replace(':', '')).get(attachURL, res => resolve(res)))
 }
 
 function sentMessageCallback (err, info) {
